% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte_cor_check.R
\name{Analyte_cor_check_res}
\alias{Analyte_cor_check_res}
\title{R6 Class representing the results from an analyte_max_cor_res call}
\description{
Used for pretty printing
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{disp}}{Display table.}

\item{\code{k}}{Number of replicates averaged.}

\item{\code{use}}{Which cross-sectional CV was used.}

\item{\code{analyte}}{The list of analytes}

\item{\code{n}}{Number of participants}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Analyte_cor_check_res-new}{\code{Analyte_cor_check_res$new()}}
\item \href{#method-Analyte_cor_check_res-print}{\code{Analyte_cor_check_res$print()}}
\item \href{#method-Analyte_cor_check_res-plot}{\code{Analyte_cor_check_res$plot()}}
\item \href{#method-Analyte_cor_check_res-clone}{\code{Analyte_cor_check_res$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte_cor_check_res-new"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte_cor_check_res-new}{}}}
\subsection{Method \code{new()}}{
Create a new Analyte_cor_check_res object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte_cor_check_res$new(disp, k, use, analyte, n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{disp}}{Display table.}

\item{\code{k}}{Number of replicates averaged.}

\item{\code{use}}{Which cross-sectional CV was used.}

\item{\code{analyte}}{List of analytes}

\item{\code{n}}{Number of participants}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Analyte_cor_check_res` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte_cor_check_res-print"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte_cor_check_res-print}{}}}
\subsection{Method \code{print()}}{
Printing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte_cor_check_res$print(digits = 3, p_digits = 3, max_rows = 20)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{The number of digits to be displayed}

\item{\code{p_digits}}{The number of digits to be displayed for p}

\item{\code{max_rows}}{The maximum number of rows to be displayed
Pretty printing of Analyte_cor_check results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte_cor_check_res-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte_cor_check_res-plot}{}}}
\subsection{Method \code{plot()}}{
Plotting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte_cor_check_res$plot(
  case = 1,
  n_points = 1000,
  xlim = c(0, 1),
  aspect.ratio = 0.5,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{case}}{The number of digits to be displayed}

\item{\code{n_points}}{The number of digits to be displayed for p}

\item{\code{xlim}}{The maximum number of rows to be displayed}

\item{\code{aspect.ratio}}{The maximum number of rows to be displayed}

\item{\code{...}}{The maximum number of rows to be displayed
Pretty printing of Analyte_cor_check results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte_cor_check_res-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte_cor_check_res-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte_cor_check_res$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
