% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte_max_cor.R
\name{Analyte_max_cor_res}
\alias{Analyte_max_cor_res}
\title{R6 Class representing the results from an analyte_max_cor_res call}
\description{
Used for pretty printing
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{disp}}{Display table.}

\item{\code{k}}{Number of replicates averaged.}

\item{\code{use}}{Which cross-sectional CV was used.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Analyte_max_cor_res-new}{\code{Analyte_max_cor_res$new()}}
\item \href{#method-Analyte_max_cor_res-print}{\code{Analyte_max_cor_res$print()}}
\item \href{#method-Analyte_max_cor_res-clone}{\code{Analyte_max_cor_res$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte_max_cor_res-new"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte_max_cor_res-new}{}}}
\subsection{Method \code{new()}}{
Create a new Analyte_max_cor_res object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte_max_cor_res$new(disp, k, use)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{disp}}{Display table.}

\item{\code{k}}{Number of replicates averaged.}

\item{\code{use}}{Which cross-sectional CV was used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Analyte_max_cor_res` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte_max_cor_res-print"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte_max_cor_res-print}{}}}
\subsection{Method \code{print()}}{
Pretty printing of analyte_max_cor results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte_max_cor_res$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Analyte_max_cor_res-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Analyte_max_cor_res-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyte_max_cor_res$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
