% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovas.R
\name{anova_one_descriptive}
\alias{anova_one_descriptive}
\title{Reconstruct results from a one-way ANOVA}
\usage{
anova_one_descriptive(m, s, n, output = FALSE)
}
\arguments{
\item{m}{Character vector. Mean for each group. \strong{Must be character} to infer rounding.}

\item{s}{Character vector. SD for each group. \strong{Must be character} to infer rounding.}

\item{n}{Numeric vector. Sample size for each group.}
}
\value{
Minimum, RIVETS and maximum p values
}
\description{
This function allows for the reconstruction of the omnibus test from a one-way ANOVA,
using the descriptive statistics for each group.
}
\examples{
anova_one_descriptive(m = c("2.51", "3.12"), s = c("0.58", "0.34"), n = c(40, 40))

}
