% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_validation.R
\name{decimal_places}
\alias{decimal_places}
\title{decimal_places}
\usage{
decimal_places(x)
}
\arguments{
\item{x}{Character string. The statistic to examine}
}
\value{
The number of decimal places
}
\description{
Finds and returns the number of decimal places on a statistic
}
\examples{
decimal_places("3")
decimal_places(".984")
decimal_places("2.00")
}
