% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{my_function}
\alias{my_function}
\title{My Function Title}
\usage{
my_function(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A numeric vector representing the sum of x and y.
}
\description{
A more detailed description of what the function does.
}
\examples{
my_function(1:3, 4:6)
}
