% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_validation.R
\name{validation_p}
\alias{validation_p}
\title{validation_descriptive}
\usage{
validation_p(p, name = "p")
}
\arguments{
\item{p}{Character string. The p value to validate}

\item{name}{Character string. Optional. The name of the variable
(for error output)}
}
\value{
A list of numbers: the possible range of values (minimum, original
and maximum)
}
\description{
Validates a p statistic: checks it is a string that can be
converted to a number, and returns the possible range of values if it was
rounded
}
\examples{
validation_p(".14")
p_value <- ".25"
validation_p(p_value, "p_value")
validation_p("<.001")
}
