% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovas.R
\docType{class}
\name{Anova_res}
\alias{Anova_res}
\title{R6 class: results of an Anova reconstruction function}
\description{
R6 class: results of an Anova reconstruction function

R6 class: results of an Anova reconstruction function
}
\details{
Object container for results produced by \code{anovas} functions.
}
\seealso{
\code{\link{anova_one_descriptive}}
}
\keyword{class}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Min}}{Minimum p value from the reconstructed
ANOVA}

\item{\code{RIVETS}}{p value from the reconstructed
ANOVA using RIVETS values}

\item{\code{Max}}{Maximum p value from the reconstructed
ANOVA}

\item{\code{Function}}{The function that was called to create those results
(DEV: in case we end up having several anova functions)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Anova_res-new}{\code{Anova_res$new()}}
\item \href{#method-Anova_res-print}{\code{Anova_res$print()}}
\item \href{#method-Anova_res-clone}{\code{Anova_res$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Anova_res-new"></a>}}
\if{latex}{\out{\hypertarget{method-Anova_res-new}{}}}
\subsection{Method \code{new()}}{
Create a new Anova_res object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Anova_res$new(Min = NULL, RIVETS = NULL, Max = NULL, Function = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Min}}{Minimum p value}

\item{\code{RIVETS}}{RIVETS p value}

\item{\code{Max}}{Max p value}

\item{\code{Function}}{The function that was called to create those results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Anova_res` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Anova_res-print"></a>}}
\if{latex}{\out{\hypertarget{method-Anova_res-print}{}}}
\subsection{Method \code{print()}}{
Pretty printing of anovas results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Anova_res$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Anova_res-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Anova_res-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Anova_res$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
