% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carlisle.R
\name{carlisle}
\alias{carlisle}
\title{Perform a Carlisle test}
\usage{
carlisle(p, output = FALSE)
}
\arguments{
\item{p}{A character vector. Either the p values from tests of differences
between the groups at baseline, or a T_slicer_res object (the results from a
t_slicer() call)}

\item{output}{A logical. Optional (default = FALSE). Determines if the
function should print its output before returning it}
}
\value{
A Carlisle_res object. Either contains a single set of min/RIVETS/max
results from the Carlisle test if p values were manually provided, or two sets
of min/RIVETS/max results (once for Student t tests, one for Welch t tests)
if a T_slicer_res object was provided.
}
\description{
A Carlisle test is a method for examining the table of participants'
baseline characteristics in reports of randomized trials, and checking if
groups are too consistently different (or similar) than what would be expected
in case of successful randomization. Takes all p values with the assumption
of independence of tests and the NULL is true to calculate an aggregated p
value using a Fisher-Stouffer method.
}
\examples{
carlisle(c(".99", ".95", ".98", ".90"), output=TRUE)
data <- tibble(m1 = c("1.20", "1.25"), s1 = c("1.2", "1.25"), n1 = c(60, 60),
               m2 = c("2.1", "2.15"), s2 = c("2.5", "2.55"), n2 = c(30, 30),
               p = c(NA, NA))
t_slicer(m1 = m1, s1 = s1, n1 = n1, m2 = m2, s2 = s2, n2 = n2,
         data = data, output = TRUE) \%>\%
         carlisle(output = TRUE)

}
