% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carlisle.R
\name{Carlisle_res}
\alias{Carlisle_res}
\title{R6 Class representing the results from a carlisle() call}
\description{
Mainly for pretty printing, but all output is available
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Student}}{Carlisle min/RIVETS/max test results using
the range of Student p values from a T_slicer_res object}

\item{\code{Welch}}{Carlisle min/RIVETS/max test results using
the range of Welch p values from a T_slicer_res object}

\item{\code{Provided_p}}{Carlisle min/RIVETS/max test results using
the range of p values manually provided}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Carlisle_res-new}{\code{Carlisle_res$new()}}
\item \href{#method-Carlisle_res-print}{\code{Carlisle_res$print()}}
\item \href{#method-Carlisle_res-clone}{\code{Carlisle_res$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Carlisle_res-new"></a>}}
\if{latex}{\out{\hypertarget{method-Carlisle_res-new}{}}}
\subsection{Method \code{new()}}{
Create a new Carlisle_res object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Carlisle_res$new(Student = NULL, Welch = NULL, Provided_p = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Student}}{Student res}

\item{\code{Welch}}{Welch res}

\item{\code{Provided_p}}{Provided_p res}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Carlisle_res` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Carlisle_res-print"></a>}}
\if{latex}{\out{\hypertarget{method-Carlisle_res-print}{}}}
\subsection{Method \code{print()}}{
Pretty printing of Carlisle results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Carlisle_res$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Carlisle_res-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Carlisle_res-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Carlisle_res$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
