% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_validation.R
\name{decimal_place}
\alias{decimal_place}
\title{Find and return the number of decimal places on a statistic}
\usage{
decimal_place(statistic)
}
\arguments{
\item{statistic}{Number or (ideally) character vectors. The statistic to
examine}
}
\value{
The number of decimal places
}
\description{
Find and return the number of decimal places on a statistic
}
\examples{
decimal_place("3")
decimal_place(".984")
decimal_place("2.00")
decimal_place(c("3.14", NA, "6.8"))
}
