% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_validation.R
\name{decimal_places}
\alias{decimal_places}
\title{Find and return the number of decimal places on a statistic}
\usage{
decimal_places(statistic)
}
\arguments{
\item{statistic}{Number or (ideally) character vectors. The statistic to
examine}
}
\value{
The number of decimal places
}
\description{
Find and return the number of decimal places on a statistic
}
\examples{
decimal_places("3")
decimal_places(".984")
decimal_places("2.00")
decimal_places(c("3.14", NA, "6.8"))
}
