% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_slicer.R
\name{T_slicer_res}
\alias{T_slicer_res}
\title{R6 Class representing the results from a t_slicer call}
\description{
Mainly for pretty printing, but all output is available
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Student_RIVETS}}{p value from Student's t test using
RIVETS stats}

\item{\code{Welch_RIVETS}}{p value from Welch's t test using
RIVETS stats}

\item{\code{Student_min}}{p value from Student's t test using
minimum possible values}

\item{\code{Student_max}}{p value from Student's t test using
maximum possible values}

\item{\code{Welch_min}}{p value from Welch's t test using
minimum possible values}

\item{\code{Welch_max}}{p value from Welch's t test using
maximum possible values}

\item{\code{Student}}{are statistics consistent with a Student's t
test?}

\item{\code{Welch}}{are statistics consistent with a Welch's t
test?}

\item{\code{Plots}}{the resulting plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-T_slicer_res-new}{\code{T_slicer_res$new()}}
\item \href{#method-T_slicer_res-print}{\code{T_slicer_res$print()}}
\item \href{#method-T_slicer_res-plot}{\code{T_slicer_res$plot()}}
\item \href{#method-T_slicer_res-clone}{\code{T_slicer_res$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-T_slicer_res-new"></a>}}
\if{latex}{\out{\hypertarget{method-T_slicer_res-new}{}}}
\subsection{Method \code{new()}}{
Create a new T_slicer_res object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{T_slicer_res$new(
  Student_RIVETS,
  Welch_RIVETS,
  Student_min,
  Student_max,
  Welch_min,
  Welch_max,
  Student,
  Welch,
  Plots
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Student_RIVETS}}{Student_RIVETS}

\item{\code{Welch_RIVETS}}{Welch_RIVETS}

\item{\code{Student_min}}{Student_min}

\item{\code{Student_max}}{Student_max}

\item{\code{Welch_min}}{Welch_min}

\item{\code{Welch_max}}{Welch_max}

\item{\code{Student}}{Student}

\item{\code{Welch}}{Welch}

\item{\code{Plots}}{Plots}

\item{\code{n}}{n}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `T_slicer_res` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-T_slicer_res-print"></a>}}
\if{latex}{\out{\hypertarget{method-T_slicer_res-print}{}}}
\subsection{Method \code{print()}}{
Pretty printing of t_slicer results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{T_slicer_res$print(n = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{if NULL (default), outputs results for all
lines. Otherwise, only the one from the specified line.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-T_slicer_res-plot"></a>}}
\if{latex}{\out{\hypertarget{method-T_slicer_res-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{T_slicer_res$plot(n = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{if NULL (default), plots a grid of all graphs for all
lines. Otherwise, plots the graph for the specified line.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-T_slicer_res-clone"></a>}}
\if{latex}{\out{\hypertarget{method-T_slicer_res-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{T_slicer_res$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
