% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_validation.R
\name{validate_descriptive}
\alias{validate_descriptive}
\title{validate_descriptive}
\usage{
validate_descriptive(statistic)
}
\arguments{
\item{statistic}{Character string. The statistic to validate}
}
\value{
A list of numbers: the possible range of values (minimum, original
and maximum)
}
\description{
Validates a descriptive statistic: checks it is a string that can be
converted to a number, and returns the possible range of values if it was
rounded
}
\examples{
validate_descriptive("2.14")
}
