% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_validation.R
\name{validate_p}
\alias{validate_p}
\title{validate_p}
\usage{
validate_p(p)
}
\arguments{
\item{p}{Character string. The p value to validate}
}
\value{
A list of numbers: the possible range of values (minimum, original
and maximum)
}
\description{
Validates a p statistic: checks it is a string that can be
converted to a number, and returns the possible range of values if it was
rounded
}
\examples{
validate_p(".14")
validate_p("<.001")
}
